Update ExecutorFacet and IExecutor to use uint64 for expiration parameters
- Changed the expiration parameter type from uint48 to uint64 in the `setGroupExpiration` function and related methods in `ExecutorFacet` and `IExecutor` to accommodate larger expiration values.
- Updated corresponding test cases to reflect the new expiration type, ensuring compatibility and correctness in expiration handling.
- Adjusted ABI and typings to align with the updated expiration parameter type, enhancing type safety across the codebase.

+++ packages/contracts/src/spaces/facets/executor/ExecutorFacet.sol
@@ -45,7 +45,7 @@ contract ExecutorFacet is OwnableBase, ExecutorBase, IExecutor, Facet {
         bytes32 groupId,
         address account,
         uint32 delay,
-        uint48 expiration
+        uint64 expiration
     ) external onlyOwner returns (bool newMember) {
         _setGroupStatus(groupId, true, expiration);
         return _grantGroupAccess(groupId, account, _getGroupGrantDelay(groupId), delay);
@@ -72,7 +72,7 @@ contract ExecutorFacet is OwnableBase, ExecutorBase, IExecutor, Facet {
     }
 
     /// @inheritdoc IExecutor
-    function setGroupExpiration(bytes32 groupId, uint48 expiration) external onlyOwner {
+    function setGroupExpiration(bytes32 groupId, uint64 expiration) external onlyOwner {
         _setGroupExpiration(groupId, expiration);
     }
